// lib/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      wedding_info: {
        Row: {
          bank_account_name: string | null
          bank_account_number: string | null
          bank_name: string | null
          bank_qr_image_url: string | null
          bride_name: string
          created_at: string | null
          groom_name: string
          hero_image_url: string | null
          id: string
          story_text: string | null
          updated_at: string | null
          venue_address: string | null
          venue_map_url: string | null
          venue_name: string | null
          wedding_date: string
          wedding_time: string | null
        }
        Insert: {
          bank_account_name?: string | null
          bank_account_number?: string | null
          bank_name?: string | null
          bank_qr_image_url?: string | null
          bride_name: string
          created_at?: string | null
          groom_name: string
          hero_image_url?: string | null
          id?: string
          story_text?: string | null
          updated_at?: string | null
          venue_address?: string | null
          venue_map_url?: string | null
          venue_name?: string | null
          wedding_date: string
          wedding_time?: string | null
        }
        Update: {
          bank_account_name?: string | null
          bank_account_number?: string | null
          bank_name?: string | null
          bank_qr_image_url?: string | null
          bride_name?: string
          created_at?: string | null
          groom_name?: string
          hero_image_url?: string | null
          id?: string
          story_text?: string | null
          updated_at?: string | null
          venue_address?: string | null
          venue_map_url?: string | null
          venue_name?: string | null
          wedding_date?: string
          wedding_time?: string | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}